feat: creates trending hashtags area
