+++ .wakatime-project
@@ -1,0 +1,1 @@
+projeto17-linkr_front-end
