--- src/HashtagsPage/index.js
@@ -1,13 +1,0 @@
-import Header from "../Header/index.js"
-import RenderHashtagsPage from "./RenderHashtagsPage.js";
-
-function HashtagsPage(){
-  return (
-    <>
-      <Header />
-      <RenderHashtagsPage />
-    </>
-  )
-};
-
-export default HashtagsPage;
